requirements.txt
# TNT T-SA Intelligence Suite - Streamlined Production Requirements
# Windows-compatible ONLY pre-built wheels (no compilation needed)

# Core ML/CV - Pre-built binaries
opencv-python==4.10.0.84
ultralytics==8.3.0
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
pillow==10.4.0

# OCR for License Plate Recognition (ANPR)
easyocr==1.7.1
paddleocr==2.7.3

# Pose Estimation for Seatbelt Detection
mediapipe==0.10.14

# FastAPI Backend
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12

# Database - SQLite only (built-in Python, no compilation)
sqlalchemy==2.0.35
aiosqlite==0.20.0
psycopg2-binary==2.9.9

# Data handling
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1

# Desktop Dashboard - PySide6 (Qt)
PySide6==6.8.0
PySide6-Addons==6.8.0
PySide6-Essentials==6.8.0
PySide6-WebEngineWidgets==6.8.0

# Async/Networking
httpx==0.27.2
requests==2.32.3

# Voice Alerts (Windows)
pyttsx3==2.98

# Logging
python-json-logger==2.0.7
colorlog==6.8.2

# Security - Pre-built wheels
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
cryptography==43.0.1

# ML Training - Using binary wheels only
pandas==2.2.3
matplotlib==3.9.2
scikit-learn==1.5.2
opencv-contrib-python==4.10.0.84

# Video Processing
imageio==2.35.1
imageio-ffmpeg==0.5.1

# Utilities
python-dateutil==2.9.0
tqdm==4.66.5
pytz==2024.2

# PDF Generation for Reports
reportlab==4.2.2

# Excel Export
openpyxl==3.1.5
xlsxwriter==3.2.0


